Init.
This code follows the patterns demonstrated by the $tcpinfo_* variables
in the core nginx code. Unfortunately it seems that the http request's
peer connection is NULL at the time that the variable data is accessed
for writing to the access log.

More work is required to implement this. Likely following the pattern
of caching the required data in the request's upstream_states member
or similar as demonstrated by the $upstream_addr variable.